feat: QPPA-0000 spring cleaning for the measures ingestion process
