+++ scripts/measures/lib/ingest-specifications-links.ts
@@ -12,15 +12,15 @@ import path from 'path';
 import { parse } from 'csv-parse/sync';
 import appRoot from 'app-root-path';
 
-import { mergeEcqmEhrLinks } from '../lib/merge-ecqm-ehr-links';
-import { mergeWebInterfaceLinks } from '../lib/merge-web-interface-links';
-import { mergeClaimsLinks } from '../lib/merge-claims-links';
-import { mergeCqmLinks } from '../lib/merge-cqm-links';
-import { mergePiLinks } from '../lib/merge-pi-links';
-import { mergeCostLinks } from '../lib/merge-cost-links';
-import { mergeEcqmData } from '../lib/merge-ecqm-data';
-import { mergeStratifications } from '../lib/merge-stratifications';
-import { mergeClaimsRelatedData } from '../lib/merge-claims-related-data';
+import { mergeEcqmEhrLinks } from './merge-ecqm-ehr-links';
+import { mergeWebInterfaceLinks } from './merge-web-interface-links';
+import { mergeClaimsLinks } from './merge-claims-links';
+import { mergeCqmLinks } from './merge-cqm-links';
+import { mergePiLinks } from './merge-pi-links';
+import { mergeCostLinks } from './merge-cost-links';
+import { mergeEcqmData } from './merge-ecqm-data';
+import { mergeStratifications } from './merge-stratifications';
+import { mergeClaimsRelatedData } from './merge-claims-related-data';
 
 const currentPerformanceYear = process.argv[2];
