+++ test/measures/2024/badPiMeasures.csv
@@ -1,0 +1,2 @@
+﻿Category,Measure ID,Measure Title,Measure Description,Required,Metric Type,First Performance Year,Year Removed,Objective,Bonus,Reporting Category,Substitutes,Exclusions
+PI,PI_PPHI_1,piTitle,piDescription,N,boolean,2018,,Public Health And Clinical Data Exchange,Yes,required,PI_PPHI_2,"PI_EP_1, PI_EP_32"
