+++ test/measures/2024/badQcdrMeasures.csv
@@ -1,0 +1,2 @@
+﻿Category,Measure Title,CMS eCQM ID,eCQM NQF,NQF,*** Quality Number (#) / QCDR #,Primary Measure Steward,Allowed QCDR Vendor ID,Measure Description,*** Measure Type  ,High Priority,First Performance Year,Year Removed,*** Collection Type(s) for Submission,Specialty Measure Sets,*** Inverse,*** Metric Type,*** Calculation Type,Collection Type(s) where Historic Benchmark Removed,Collection Type(s) where Suppressed,Collection Type(s) where Truncated,Is Risk Adjusted 
+QCDR,qcdrTitle,,idX11,60,TestQCDR,stewardTitle,"123456, 654321",qcdrDescription,Process,N,2018,2023,"QCDR, Part B Claims, MIPS CQM",,N,multiPerformanceRate,simpleAverage,"MIPS CQM, QCDR",MIPS CQM,Part Baddata Claims,Y
