+++ test/measures/2024/iaMeasures.csv
@@ -1,0 +1,4 @@
+﻿Category,Measure Title,Measure Description,Measure ID,First Performance Year,Year Removed,Weight,Subcategory Name
+Value: IA,,,,,Enter current PY if measure is to be removed,"Values: high, medium, NULL","Values:  Achieving Health Equity, Behavioral And Mental Health, Beneficiary Engagement, Care Coordination, Emergency Response And Preparedness, Expanded Practice Access, Patient Safety And Practice Assessment, Population Management, NULL"
+IA,iaTitle,iaDescription,IA_EPA_2,2018,,high,Population Management
+IA,iaTitle,iaDescription,IA_EPA_3,2018,,NULL,NULL
