+++ test/measures/2024/piMeasures.csv
@@ -1,0 +1,3 @@
+﻿Category,Measure ID,Measure Title,Measure Description,Required,Metric Type,First Performance Year,Year Removed,Objective,Bonus,Reporting Category,Substitutes,Exclusions
+PI,PI_PPHI_1,piTitle,piDescription,FALSE,boolean,2018,,Public Health And Clinical Data Exchange,TRUE,required,PI_PPHI_2,"PI_EP_1, PI_EP_32"
+PI,PI_PPHI_2,piTitle,piDescription,FALSE,boolean,2018,,NULL,TRUE,NULL,NULL,NULL
