+++ test/measures/2024/quality-strata.csv
@@ -1,0 +1,16 @@
+measureId,stratumName,description
+009,>=84Days,Percentage of patients who remained on an antidepressant medication for at least 84 days (12 weeks).
+009,>=180Days,Percentage of patients who remained on an antidepressant medication for at least 180 days (6 months)
+226,screenedForUse,Percentage of patients aged 18 years and older who were screened for tobacco use one or more times within the measurement period
+226,overall,Percentage of patients aged 18 years and older who were identified as a tobacco user who received tobacco cessation intervention on the date of the encounter or within the previous 12 months
+226,tobacco,Percentage of patients aged 18 years and older who were screened for tobacco use one or more times within the measurement period AND who received tobacco cessation intervention if identified as a tobacco user on the date of the encounter or within the previous 12 months
+238,overall,Percentage of patients 65 years of age and older who were ordered at least two high-risk medications from the same drug class.
+238,diagnosisException,"Percentage of patients 65 years of age and older who were ordered at least two high-risk medications from the same drug class, except for appropriate diagnoses."
+238,eCQMTotalRate,"Total rate (the sum of the two numerators divided by the denominator, deduplicating for patients in both numerators)."
+239,BMI,"Patients who had a height, weight and body mass index (BMI) percentile recorded during the measurement period"
+239,nutrition,Patients who had counseling for nutrition during a visit that occurs during the measurement period
+239,physicalActivity,Patients who had counseling for physical activity during a visit that occurs during the measurement period
+305,14DaysOfDiagnosis,Patients who initiated treatment within 14 days of the diagnosis
+305,30DaysOfVisit,Patients who initiated treatment and who had two or more additional services with an AOD diagnosis within 30 days of the initiation visit
+366,visitWithin30Days,Percentage of children who had one follow-up visit with a practitioner with prescribing authority during the 30-Day Initiation Phase.
+366,overall,"Percentage of children who remained on ADHD medication for at least 210 days and who, in addition to the visit in the Initiation Phase, had at least two additional follow-up visits with a practitioner within 270 days (9 months) after the Initiation Phase ended."
