+++ test/measures/2024/qualityMeasures.csv
@@ -1,0 +1,3 @@
+﻿Category,Measure Title,CMS eCQM ID,eCQM NQF,NQF,*** Quality Number (#) / QCDR #,Primary Measure Steward,Allowed QCDR Vendor ID,Measure Description,*** Measure Type  ,High Priority,First Performance Year,Year Removed,*** Collection Type(s) for Submission,Specialty Measure Sets,*** Inverse,*** Metric Type,*** Calculation Type,Collection Type(s) where Historic Benchmark Removed,Collection Type(s) where Suppressed,Collection Type(s) where Truncated,Is Risk Adjusted 
+Quality,qualityTitle,,idX11,0060,001,stewardTitle,,qualityDescription,Process,N,2018,,"QCDR, Part B Claims, MIPS CQM","Nephrology, Preventive Medicine",N,singlePerformanceRate,simpleAverage,"MIPS CQM, QCDR",MIPS CQM,Part B Claims,Y
+QCDR,qcdrTitle,,idX11,0060,TestQCDR,stewardTitle,"123456, 654321",qcdrDescription,Process,N,2018,,"QCDR, Part B Claims, MIPS CQM",,N,singlePerformanceRate,simpleAverage,"MIPS CQM, QCDR",MIPS CQM,Part B Claims,Y
